requirements.txt
certifi
chardet
django-crispy-forms
idna
Pillow
pytz
requests
urllib3
